Address comments and move common protobuf logic into src/protobuf
