+++ src/protobuf/mod.rs
@@ -1,0 +1,11 @@
+mod descriptor;
+mod encode;
+mod parse;
+
+pub use descriptor::get_message_descriptor;
+
+pub use encode::encode_message;
+pub(crate) use encode::encode_proto;
+
+pub(crate) use parse::parse_proto;
+pub use parse::proto_to_value;
